RoomManager: Add abstract RoomManager
This class will be extended by atleast 2 implementations:
    - A legacy implementation using createGroup()
    - A wifi direct implementation using connect()

The former creates a room with current device as the group owner
and the latter decides the group owner based on battey power, signal
strength etc.

For now, legacy implementation will have higher priority so that we
can assign the group owner ourselves.